MindSpore/YoloV5/requirements.txt
pandas
numpy==1.23.4
Pillow==9.3.0
